requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2023.11.17
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2023.12.2
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
pillow==10.2.0
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
sympy==1.12
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
